psi: Split up psi.go, add sample.conf and README
