+++ plugins/inputs/psi/sample.conf
@@ -1,0 +1,4 @@
+# Read metrics about Pressure Stall Information (PSI)
+# This plugin ONLY supports Linux v4.20+
+[[inputs.psi]]
+  # no configuration
